* Update input files.    plume.in
* Begin removing src directories.  Source code is being moved up one level.
  Now modules will be only one directory deep so that, for example, what
  was in bing/src is now in bing.  Extra module data such as sample input
  files and matlab files will be moved to new location (as yet to determined).
      * failure
      * sedutils
      * libsed (also changed name to sed)
      * quake
      * xshore
      * libutils (also changed name to utils)
      * subside
      * avulsion

* Updated standalone plume model.  Input files now reflect those of sakura.

* Mass balance in plumeout2.c.  The final deposit was scaled so that the output
  mass matched the input mass.  However, it is possible to lose mass out the end
  of the domain.  In such a case, the output mass should not match the input mass
  and so we do not want to scale for mass balance.  Thus, we NO LONGER scale to
  ensure mass balance.  We may want to revisit this if we begin to see large
  errors.

* sakura.  Fixed bugs in sakura and updated user interface for standalone version.
  This seems to be working well now.  Tested that standalone version matches the
  version running within sedflux.

* Added new functionality to sedflux-read-hydro.  Now concats files, and outputs
  in the 'inline' format for sedflux.

* Cleaned up squall directory.  Trashed a bunch of old autoconf/make stuff.

* sed_diag.[ch] as a means to diagnose mass balance errors in sedflux.  A Sed_diag
  object is created and monitored to track mass balance in a Sed_cell, Sed_column,
  or Sed_cube.

* sed_cube.c: sed_cube_sediment_mass to calculate the mass of SEDIMENT in a cube.
  This does not include the mass of water, just sediment.  Use sed_cube_mass to
  get the mass of sediment plus water.  Note though that if, for example, a cell
  is compacted, its mass will change as it has lost water (but not sediment).

* sed_cell.c: sed_cell_new_bedload to create a cell that contains only bedload.

* sed_river.c: sed_river_concentration to calculate the sediment concentration of
  a river.

* sed_hydrotrend.c:
     * sed_hydrotrend_error_quark for error handling.
     * sed_hydrotrend_join_header_from_byte_order to join a series of HydroTrend
       headers into a single header.
     * sed_hydrotrend_header_destroy to destroy a hydrotrend header object.

* sed_hydro.c:
     * sed_hydro_fprint_rec to print a hydro record to standard sedflux 'inline'
       format.
     * sed_hydro_eventize_conc to eventize an array of hydro records based on
       sediment concentration rather than total sediment load.

* sed headers are now installed in $(include_dir)/sed

* compact: cleaned up code.  The standalone version now uses standard command line
  option syntax.

* eh_key_file.c: eh_key_file_get_int_value to get an integer value from a key-file.

* eh_str.c: eh_str_find to match a given string to the first occurence in an array
  of strings.  This is CASE INSENSITIVE.

* utils headers are now installed in $(include_dir)/utils

* sedflux/run_bedload.c: Cleaned up the code quite a lot.  Fixed some mass balance
  problems.  Changed input.  The input now is the fraction of the sediment that is
  retained on the delta plain (rather than the ratio of flood plain to ocean).

+++ sedflux/src/run_storm.c
@@ -23,7 +23,9 @@
 
 #include <stdio.h>
 #include "my_processes.h"
-#include "eh_rand.h"
+#include <utils/eh_rand.h>
+
+#include "sedflux.h"
 
 typedef struct
 {
@@ -97,8 +99,8 @@ run_storm( Sed_process proc , Sed_cube prof )
    {
       GSList *this_link;
       Sed_ocean_storm this_storm;
-      gssize n = g_slist_length( storm_list );
-      gssize i = 0;
+      gint   n = g_slist_length( storm_list );
+      gint   i = 0;
       double this_time = start_time;
       double significant_storm = G_MINDOUBLE;
